+++ app/Http/Controllers/DashboardController.php
@@ -9,10 +9,10 @@
  * Licensed under GNU General Public License 3.0.
  * Some rights reserved. See LICENSE, AUTHORS.
  *
- *  @author      Paco Orozco <paco@pacoorozco.info>
- *  @copyright   2016 Paco Orozco
- *  @license     GPL-3.0 <http://spdx.org/licenses/GPL-3.0>
- *  @link        https://github.com/pacoorozco/probind
+ * @author      Paco Orozco <paco@pacoorozco.info>
+ * @copyright   2016 Paco Orozco
+ * @license     GPL-3.0 <http://spdx.org/licenses/GPL-3.0>
+ * @link        https://github.com/pacoorozco/probind
  *
  */
 
@@ -22,6 +22,9 @@ use App\Record;
 use App\Server;
 use App\User;
 use App\Zone;
+use Carbon\Carbon;
+use Illuminate\Support\Collection;
+use Spatie\Activitylog\Models\Activity;
 
 class DashboardController extends Controller
 {
@@ -29,7 +32,7 @@ class DashboardController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\View\View
      */
     public function index()
     {
@@ -39,7 +42,10 @@ class DashboardController extends Controller
         $data['records'] = Record::all()->count();
         $data['users'] = User::all()->count();
 
+        $activityLog = Activity::orderBy('created_at', 'desc')->simplePaginate(10);
+
         return view('dashboard.index')
-            ->with('data', $data);
+            ->with('data', $data)
+            ->with('activityLog', $activityLog);
     }
 }