+++ config/app.php
@@ -119,7 +119,7 @@ return [
     |
     */
 
-    'version' => '0.2.1',
+    'version' => '0.2.2',
 
     /*
     |--------------------------------------------------------------------------
@@ -165,6 +165,7 @@ return [
         Collective\Html\HtmlServiceProvider::class, // HTML helpers
         Yajra\Datatables\DatatablesServiceProvider::class, // DataTables helpers
         Torann\Registry\RegistryServiceProvider::class, // Settings helpers
+        Spatie\Activitylog\ActivitylogServiceProvider::class, // Activity Log helpers
 
         /*
          * Application Service Providers...